+++ sperl_package.h
@@ -8,9 +8,9 @@ struct SPerl_package {
   SPerl_OP* op_type;
   SPerl_ARRAY* op_fields;
   SPerl_ARRAY* op_subs;
-  int64_t byte_size;
-  int64_t constant_pool_address;
-  size_t name_constant_pool_address;
+  int32_t byte_size;
+  int32_t constant_pool_address;
+  int32_t name_constant_pool_address;
 };
 
 SPerl_PACKAGE* SPerl_PACKAGE_new(SPerl* sperl);
